Rename project list
